+++ package.json
@@ -33,6 +33,7 @@
     "react-day-picker": "8.10.1",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.53.2",
+    "react-router-dom": "^7.3.0",
     "tailwind-merge": "^2.5.5",
     "tailwindcss-animate": "^1.0.7",
     "vite-plugin-env-compatible": "^2.0.1",
